compliance-platform-opensource/requirements.txt
flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
fpdf2==2.7.5
requests==2.31.0
google-generativeai==0.3.0
python-dotenv==1.0.0
